# frontend-angular/simulador-logistico-web/src/app/api/model/apiError.ts
/**
 * E-Cargo Hub API
 *
 * Contact: admin@simulador-logistico.local
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApiError { 
    timestamp: string;
    status: number;
    error: string;
    message: string;
    path: string;
}