# frontend-angular/simulador-logistico-web/src/app/api/model/createVehicleRequest.ts
/**
 * E-Cargo Hub API
 *
 * Contact: admin@simulador-logistico.local
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { VehicleType } from './vehicleType';


export interface CreateVehicleRequest { 
    name: string;
    type: VehicleType;
    maxSpeed: number;
}
export namespace CreateVehicleRequest {
}