# frontend-angular/simulador-logistico-web/src/app/api/model/vehicle.ts
/**
 * E-Cargo Hub API
 *
 * Contact: admin@simulador-logistico.local
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { VehiclePosition } from './vehiclePosition';
import { VehicleType } from './vehicleType';
import { VehicleStatus } from './vehicleStatus';


export interface Vehicle { 
    id: number;
    name: string;
    type: VehicleType;
    maxSpeed: number;
    status: VehicleStatus;
    currentPosition?: VehiclePosition | null;
}
export namespace Vehicle {
}