# frontend-angular/simulador-logistico-web/src/app/api/model/vehicleStatusDto.ts
/**
 * E-Cargo Hub API
 *
 * Contact: admin@simulador-logistico.local
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface VehicleStatusDto { 
    vehicleId: number;
    latitude?: number | null;
    longitude?: number | null;
    progress: number;
    speedKmh: number;
    /**
     * Estado actual del vehículo
     */
    status: string;
    updatedAt?: string | null;
}