# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: E-Cargo Hub API
  version: 1.0.0
  description: |
    El sistema permite:
      - Gestionar vehículos.
      - Gestionar almacenes.
      - Crear rutas.
      - Crear y controlar simulaciones.
      - Enviar comandos a vehículos.
      - Consultar posiciones.
      - Gestionar alertas.
      - Integrarse con GraphHopper y OpenStreetMap.
      - Distribuir eventos en tiempo real mediante Kafka.
      - Exponer actualizaciones al frontend mediante WebSocket.

  contact:
    name: Simulador Logístico
    email: admin@simulador-logistico.local
servers:
  - url: http://localhost:8081
    description: Desarrollo local
...
paths:
...
  /api/v1/vehicles:

    get:
      tags:
        - Vehicles

      summary: Obtener vehículos

      operationId: getVehicles

      responses:

        '200':
          description: Lista de vehículos

          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'

        '500':
          $ref: '#/components/responses/InternalServerError'


    post:
      tags:
        - Vehicles

      summary: Crear vehículo

      operationId: createVehicle

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleRequest'

      responses:

        '201':
          description: Vehículo creado

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'

        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/vehicles/{vehicleId}:

    get:
      tags:
        - Vehicles

      summary: Obtener vehículo

      operationId: getVehicle

      parameters:
        - $ref: '#/components/parameters/VehicleId'

      responses:

        '200':
          description: Vehículo

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'

        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Vehicles

      summary: Actualizar vehículo

      operationId: updateVehicle

      parameters:
        - $ref: '#/components/parameters/VehicleId'

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleRequest'

      responses:

        '200':
          description: Vehículo actualizado

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'

        '400':
          $ref: '#/components/responses/BadRequest'

        '404':
          $ref: '#/components/responses/NotFound'


    delete:
      tags:
        - Vehicles

      summary: Eliminar vehículo

      operationId: deleteVehicle

      parameters:
        - $ref: '#/components/parameters/VehicleId'

      responses:

        '204':
          description: Vehículo eliminado

        '404':
          $ref: '#/components/responses/NotFound'


  # ============================================================
  # WAREHOUSES
  # ============================================================
...
  /api/v1/simulations:

    get:
      tags:
        - Simulations

      summary: Obtener simulaciones

      operationId: getSimulations

      responses:

        '200':
          description: Lista de simulaciones

          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Simulation'


    post:
      tags:
        - Simulations

      summary: Crear simulación

      operationId: createSimulation

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSimulationRequest'

      responses:

        '201':
          description: Simulación creada

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Simulation'

        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/simulations/{simulationId}:

    get:
      tags:
        - Simulations

      summary: Obtener simulación

      operationId: getSimulation

      parameters:
        - $ref: '#/components/parameters/SimulationId'

      responses:

        '200':
          description: Simulación

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Simulation'

        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/simulations/{simulationId}/start:

    post:
      tags:
        - Simulations

      summary: Iniciar simulación

      operationId: startSimulation

      parameters:
        - $ref: '#/components/parameters/SimulationId'

      responses:

        '200':
          description: Simulación iniciada

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Simulation'

        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/simulations/{simulationId}/pause:

    post:
      tags:
        - Simulations

      summary: Pausar simulación

      operationId: pauseSimulation

      parameters:
        - $ref: '#/components/parameters/SimulationId'

      responses:

        '200':
          description: Simulación pausada

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Simulation'
  /api/v1/simulations/{simulationId}/resume:

    post:
      tags:
        - Simulations

      summary: Reanudar simulación

      operationId: resumeSimulation

      parameters:
        - $ref: '#/components/parameters/SimulationId'

      responses:

        '200':
          description: Simulación reanudada

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Simulation'
  /api/v1/simulations/{simulationId}/stop:

    post:
      tags:
        - Simulations

      summary: Detener simulación

      operationId: stopSimulation

      parameters:
        - $ref: '#/components/parameters/SimulationId'

      responses:

        '200':
          description: Simulación detenida

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Simulation'


  # ============================================================
  # VEHICLE COMMANDS
  # ============================================================
...
  /api/v1/vehicles/{vehicleId}/status:
    get:
      tags:
        - Vehicles

      summary: Obtener el estado actual del vehículo

      operationId: getVehicleStatus

      parameters:
        - $ref: '#/components/parameters/VehicleId'

      responses:
        '200':
          description: Estado actual, o estado IDLE si aún no hay datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleStatusDto'
  /api/v1/vehicles/{vehicleId}/telemetry:

    get:
      tags:
        - Telemetry

      summary: Obtener histórico completo de telemetría

      operationId: getVehicleTelemetry

      description: |
        Devuelve todos los puntos de telemetría registrados para el vehículo,
        ordenados cronológicamente. Se puede filtrar opcionalmente desde, hasta,
        o por ambos límites de fecha.

      parameters:
        - $ref: '#/components/parameters/VehicleId'

        - name: from
          in: query
          required: false
          description: Fecha de inicio (ISO-8601)
          schema:
            type: string
            format: date-time

        - name: to
          in: query
          required: false
          description: Fecha de fin (ISO-8601)
          schema:
            type: string
            format: date-time

      responses:

        '200':
          description: Lista de puntos de telemetría

          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetryPoint'

    delete:
      tags:
        - Telemetry

      summary: Borrar el histórico de telemetría

      operationId: clearVehicleTelemetry

      parameters:
        - $ref: '#/components/parameters/VehicleId'

      responses:
        '200':
          description: Número de puntos eliminados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetryClearResult'
  /api/v1/vehicles/{vehicleId}/telemetry/latest:

    get:
      tags:
        - Telemetry

      summary: Obtener los últimos N puntos de telemetría

      operationId: getLatestVehicleTelemetry

      parameters:
        - $ref: '#/components/parameters/VehicleId'

        - name: limit
          in: query
          required: false
          description: Número de puntos a devolver (entre 1 y 500)
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 500

      responses:

        '200':
          description: Lista de puntos recientes

          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetryPoint'



  # ============================================================
  # ALERTS
  # ============================================================
...
components:
...
  parameters:
    VehicleId:
      name: vehicleId
      in: path
      required: true
      description: Identificador del vehículo
      schema:
        type: integer
        format: int64
        minimum: 1
...
    SimulationId:
      name: simulationId
      in: path
      required: true
      description: Identificador de la simulación
      schema:
        type: integer
        format: int64
        minimum: 1


  # ============================================================
  # SCHEMAS
  # ============================================================
  schemas:
...
    Vehicle:
      type: object

      required:
        - id
        - name
        - type
        - maxSpeed
        - status

      properties:

        id:
          type: integer
          format: int64

        name:
          type: string
          example: TRUCK-001

        type:
          $ref: '#/components/schemas/VehicleType'

        maxSpeed:
          type: number
          format: double
          minimum: 0
          example: 120

        status:
          $ref: '#/components/schemas/VehicleStatus'

        currentPosition:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/VehiclePosition'
    CreateVehicleRequest:
      type: object

      required:
        - name
        - type
        - maxSpeed

      properties:

        name:
          type: string
          minLength: 1
          maxLength: 100
          example: TRUCK-001

        type:
          $ref: '#/components/schemas/VehicleType'

        maxSpeed:
          type: number
          format: double
          minimum: 0
          example: 120
    UpdateVehicleRequest:
      type: object

      required:
        - name
        - type
        - maxSpeed

      properties:

        name:
          type: string
          minLength: 1
          maxLength: 100

        type:
          $ref: '#/components/schemas/VehicleType'

        maxSpeed:
          type: number
          format: double
          minimum: 0
    VehicleType:
      type: string

      enum:
        - CAR
        - VAN
        - TRUCK
    VehicleStatus:
      type: string

      enum:
        - IDLE
        - RUNNING
        - PAUSED
        - STOPPED
        - ERROR
    VehicleStatusDto:
      type: object

      required:
        - vehicleId
        - progress
        - speedKmh
        - status

      properties:

        vehicleId:
          type: integer
          format: int64

        latitude:
          type: number
          format: double
          nullable: true
          minimum: -90
          maximum: 90

        longitude:
          type: number
          format: double
          nullable: true
          minimum: -180
          maximum: 180

        progress:
          type: integer
          minimum: 0
          maximum: 100

        speedKmh:
          type: number
          format: double
          minimum: 0

        status:
          type: string
          description: Estado actual del vehículo
          example: IDLE

        updatedAt:
          type: string
          format: date-time
          nullable: true


    # ------------------------------------------------------------
    # WAREHOUSE
    # ------------------------------------------------------------
...
    Simulation:
      type: object

      required:
        - id
        - name
        - status
        - routeId
        - vehicleId

      properties:

        id:
          type: integer
          format: int64

        name:
          type: string
          example: Simulación Alicante-Madrid

        vehicleId:
          type: integer
          format: int64

        routeId:
          type: integer
          format: int64

        status:
          $ref: '#/components/schemas/SimulationStatus'

        startedAt:
          type: string
          format: date-time
          nullable: true

        finishedAt:
          type: string
          format: date-time
          nullable: true

        simulationSpeed:
          type: number
          format: double
          nullable: true
          minimum: 0
          example: 1
    CreateSimulationRequest:
      type: object

      required:
        - name
        - vehicleId
        - routeId

      properties:

        name:
          type: string
          minLength: 1
          maxLength: 150

        vehicleId:
          type: integer
          format: int64

        routeId:
          type: integer
          format: int64

        simulationSpeed:
          type: number
          format: double
          minimum: 0
          default: 1
    SimulationStatus:
      type: string

      enum:
        - CREATED
        - RUNNING
        - PAUSED
        - FINISHED
        - STOPPED
        - ERROR


    # ------------------------------------------------------------
    # VEHICLE POSITION
    # ------------------------------------------------------------
    VehiclePosition:
      type: object

      required:
        - vehicleId
        - latitude
        - longitude
        - speed
        - heading
        - timestamp

      properties:

        vehicleId:
          type: integer
          format: int64

        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          example: 38.3452

        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          example: -0.4810

        speed:
          type: number
          format: double
          minimum: 0
          example: 92.5

        speedLimit:
          type: number
          format: double
          minimum: 0
          example: 90

        acceleration:
          type: number
          format: double
          example: 0.8

        heading:
          type: number
          format: double
          minimum: 0
          maximum: 360
          example: 274.2

        timestamp:
          type: string
          format: date-time
    TelemetryPoint:
      type: object

      required:
        - id
        - vehicleId
        - latitude
        - longitude
        - progress
        - speedKmh
        - status
        - recordedAt

      properties:

        id:
          type: integer
          format: int64
          example: 1234

        vehicleId:
          type: integer
          format: int64
          example: 1

        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          example: 38.3452

        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          example: -0.4810

        progress:
          type: integer
          minimum: 0
          maximum: 100
          example: 45

        speedKmh:
          type: number
          format: double
          minimum: 0
          example: 87.5

        status:
          type: string
          example: EN_RUTA

        recordedAt:
          type: string
          format: date-time
          example: '2026-09-30T14:30:00Z'
    TelemetryClearResult:
      type: object
      required:
        - vehicleId
        - deleted
      properties:
        vehicleId:
          type: integer
          format: int64
        deleted:
          type: integer
          format: int64

    # ------------------------------------------------------------
    # COMMANDS
    # ------------------------------------------------------------
...
    ApiError:
      type: object

      required:
        - timestamp
        - status
        - error
        - message
        - path

      properties:

        timestamp:
          type: string
          format: date-time

        status:
          type: integer
          example: 404

        error:
          type: string
          example: NOT_FOUND

        message:
          type: string
          example: Vehicle not found

        path:
          type: string
          example: /api/v1/vehicles/10


  # ============================================================
  # RESPONSES
  # ============================================================
  responses:
    BadRequest:
      description: Petición incorrecta

      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Recurso no encontrado

      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    InternalServerError:
      description: Error interno del servidor

      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'